# docs/app/swagger.yaml
basePath: /api/v1
definitions:
...
  placio-app_Dto.BusinessDto:
    properties:
      cover_image:
        type: string
      description:
        type: string
      email:
        type: string
      location:
        type: string
      name:
        type: string
      phone:
        type: string
      picture:
        type: string
      website:
        type: string
    type: object
...
  placio-app_Dto.Error:
    properties:
      message:
        type: string
      status:
        type: integer
    type: object
  placio-app_Dto.ErrorDTO:
    properties:
      message:
        type: string
      status:
        type: integer
    type: object
  placio-app_Dto.ErrorDto:
    properties:
      message:
        type: string
      status:
        type: integer
    type: object
...
  placio-app_Dto.Response:
    properties:
      data: {}
      message:
        type: string
      status:
        type: integer
    type: object
...
host: https://api.palnight.com
info:
  contact:
    email: support@placio.io
    name: Placio Ltd
    url: https://placio.io
  description: This is the documentation for the Placio Application Api
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: https://placio.io/terms
  title: Placio Application Api
  version: "0.01"
paths:
...
  /api/v1/business/:
    patch:
      consumes:
      - application/json
      description: Get a business's details by ID
      parameters:
      - description: Business ID
        in: path
        name: id
        required: true
        type: string
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved business
          schema:
            $ref: '#/definitions/placio-app_ent.Business'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/placio-app_Dto.ErrorDTO'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/placio-app_Dto.ErrorDTO'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/placio-app_Dto.ErrorDTO'
      summary: Update a business's details
      tags:
      - Business
...
  /api/v1/users/{followerID}/follow/business/{businessID}:
    post:
      consumes:
      - application/json
      description: Follow a business by its ID
      parameters:
      - description: ID of the follower
        in: path
        name: followerID
        required: true
        type: string
      - description: ID of the business to follow
        in: path
        name: businessID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully followed business
          schema:
            $ref: '#/definitions/placio-app_Dto.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/placio-app_Dto.ErrorDTO'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/placio-app_Dto.ErrorDTO'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/placio-app_Dto.ErrorDTO'
      summary: Follow a business
      tags:
      - User
  /api/v1/users/{followerID}/follow/user/{followedID}:
    post:
      consumes:
      - application/json
      description: Follow a user by their ID
      parameters:
      - description: ID of the follower
        in: path
        name: followerID
        required: true
        type: string
      - description: ID of the user to follow
        in: path
        name: followedID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully followed user
          schema:
            $ref: '#/definitions/placio-app_Dto.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/placio-app_Dto.ErrorDTO'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/placio-app_Dto.ErrorDTO'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/placio-app_Dto.ErrorDTO'
      summary: Follow a user
      tags:
      - User
...
  /business/:
    get:
      operationId: list-business-accounts
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/placio-app_Dto.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/placio-app_Dto.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/placio-app_Dto.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/placio-app_Dto.ErrorDto'
      summary: List all business accounts
      tags:
      - Business
    post:
      consumes:
      - application/json
      operationId: create-business-account
      parameters:
      - description: Business Account Data
        in: body
        name: Dto.BusinessDto
        required: true
        schema:
          $ref: '#/definitions/placio-app_Dto.BusinessDto'
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/placio-app_Dto.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/placio-app_Dto.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/placio-app_Dto.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/placio-app_Dto.ErrorDto'
      summary: Create a business account
      tags:
      - Business
  /business/{businessAccountID}:
    delete:
      operationId: delete-business-account
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Business Account ID
        in: path
        name: businessAccountID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/placio-app_Dto.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/placio-app_Dto.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/placio-app_Dto.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/placio-app_Dto.ErrorDto'
      summary: Delete a business account
      tags:
      - Business
    get:
      operationId: get-business-account
      parameters:
      - description: Business Account ID
        in: path
        name: businessAccountID
        required: true
        type: string
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/placio-app_Dto.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/placio-app_Dto.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/placio-app_Dto.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/placio-app_Dto.ErrorDto'
      summary: Get a business account
      tags:
      - Business
...
  /business/{businessAccountID}/event:
    post:
      consumes:
      - application/json
      description: Add a new Event to a Business Account
      operationId: add-a-new-event-to-business-account
      parameters:
      - description: Business Account ID
        in: path
        name: businessAccountID
        required: true
        type: string
      - description: Event DTO
        in: body
        name: Dto.EventDTO
        required: true
        schema:
          $ref: '#/definitions/placio-app_Dto.EventDTO'
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/placio-app_ent.Event'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/placio-app_Dto.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/placio-app_Dto.Error'
      summary: Add a new Event to a Business Account
      tags:
      - Business
...
  /business/{businessAccountID}/user/{userID}:
    delete:
      operationId: remove-user-business-account
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Business Account ID
        in: path
        name: businessAccountID
        required: true
        type: string
      - description: User ID
        in: path
        name: userID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/placio-app_Dto.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/placio-app_Dto.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/placio-app_Dto.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/placio-app_Dto.ErrorDto'
      summary: Remove user from business account
      tags:
      - Business
    post:
      operationId: associate-user-business-account
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Business Account ID
        in: path
        name: businessAccountID
        required: true
        type: string
      - description: User ID
        in: path
        name: userID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/placio-app_Dto.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/placio-app_Dto.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/placio-app_Dto.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/placio-app_Dto.ErrorDto'
      summary: Associate user with business account
      tags:
      - Business
...
  /business/{businessID}/follow/user/{userID}:
    post:
      consumes:
      - application/json
      description: Retrieve all comments for a given post
      operationId: follow-user
      parameters:
      - description: Business ID
        in: path
        name: businessID
        required: true
        type: string
      - description: User ID
        in: path
        name: userID
        required: true
        type: string
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/placio-app_Dto.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/placio-app_Dto.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/placio-app_Dto.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/placio-app_Dto.ErrorDto'
      summary: Follow a user by a business
      tags:
      - Business
...
  /business/{businessID}/unfollow/user/{userID}:
    delete:
      operationId: unfollow-user
      parameters:
      - description: Business ID
        in: path
        name: businessID
        required: true
        type: string
      - description: User ID
        in: path
        name: userID
        required: true
        type: string
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/placio-app_Dto.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/placio-app_Dto.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/placio-app_Dto.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/placio-app_Dto.ErrorDto'
      summary: Unfollow a user by a business
      tags:
      - Business
  /business/{followerBusinessID}/follow/business/{followedID}:
    post:
      operationId: follow-business
      parameters:
      - description: Follower Business ID
        in: path
        name: followerBusinessID
        required: true
        type: string
      - description: Followed Business ID
        in: path
        name: followedBusinessID
        required: true
        type: string
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/placio-app_Dto.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/placio-app_Dto.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/placio-app_Dto.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/placio-app_Dto.ErrorDto'
      summary: Follow a business by another business
      tags:
      - Business
  /business/{followerBusinessID}/unfollow/business/{followedBusinessID}:
    delete:
      operationId: unfollow-business
      parameters:
      - description: Follower Business ID
        in: path
        name: followerBusinessID
        required: true
        type: string
      - description: Followed Business ID
        in: path
        name: followedBusinessID
        required: true
        type: string
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/placio-app_Dto.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/placio-app_Dto.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/placio-app_Dto.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/placio-app_Dto.ErrorDto'
      summary: Unfollow a business by another business
      tags:
      - Business
...
  /follow/business/{businessID}:
    post:
      consumes:
      - application/json
      description: Follow a specific Business by the user
      operationId: user-follow-business
      parameters:
      - description: Business ID
        in: path
        name: businessID
        required: true
        type: string
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: User successfully followed the business
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/placio-app_Dto.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/placio-app_Dto.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/placio-app_Dto.Error'
      summary: Follow a business
      tags:
      - Follow
  /follow/event/{eventID}:
    post:
      consumes:
      - application/json
      description: Follow a specific Event by the user
      operationId: follow-event
      parameters:
      - description: Event ID
        in: path
        name: eventID
        required: true
        type: string
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: User successfully followed the event
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/placio-app_Dto.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/placio-app_Dto.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/placio-app_Dto.Error'
      summary: Follow an event
      tags:
      - Follow
  /follow/place/{placeID}:
    post:
      consumes:
      - application/json
      description: Follow a specific Place by the user
      operationId: follow-place
      parameters:
      - description: Place ID
        in: path
        name: placeID
        required: true
        type: string
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: User successfully followed the place
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/placio-app_Dto.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/placio-app_Dto.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/placio-app_Dto.Error'
      summary: Follow a place
      tags:
      - Follow
  /follow/user/{userID}:
    post:
      consumes:
      - application/json
      description: Follow a specific User
      operationId: user-follow-user
      parameters:
      - description: User ID
        in: path
        name: userID
        required: true
        type: string
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: User successfully followed the user
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/placio-app_Dto.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/placio-app_Dto.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/placio-app_Dto.Error'
      summary: Follow a user
      tags:
      - Follow
  /followed/businesses:
    get:
      consumes:
      - application/json
      description: Get all Businesses followed by the user
      operationId: get-followed-businesses
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/placio-app_ent.Business'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/placio-app_Dto.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/placio-app_Dto.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/placio-app_Dto.Error'
      summary: Get followed businesses by a user
      tags:
      - Follow
  /followed/events:
    get:
      consumes:
      - application/json
      description: Get all Events followed by the user
      operationId: get-followed-events
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/placio-app_ent.Event'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/placio-app_Dto.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/placio-app_Dto.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/placio-app_Dto.Error'
      summary: Get followed events by a user
      tags:
      - Follow
  /followed/places:
    get:
      consumes:
      - application/json
      description: Get all Places followed by the user
      operationId: get-followed-places
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/placio-app_ent.Place'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/placio-app_Dto.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/placio-app_Dto.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/placio-app_Dto.Error'
      summary: Get followed places by a user
      tags:
      - Follow
  /followed/users:
    get:
      consumes:
      - application/json
      description: Get all Users followed by the user
      operationId: get-followed-users
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/placio-app_ent.User'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/placio-app_Dto.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/placio-app_Dto.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/placio-app_Dto.Error'
      summary: Get followed users by a user
      tags:
      - Follow
...
schemes:
- http
- https
swagger: "2.0"